refactor: wraps feature into `components`, `messages`, and `validators` folder
