+++ src/bot/events/interaction-create/embed/role-grant/create/validators/index.ts
@@ -1,9 +1,9 @@
 import { ChannelType, Guild, GuildMember, Interaction, PermissionFlagsBits, Role, TextChannel } from "discord.js"
-import { RoleGrantModalError } from "./modal"
-import { ERR, MSG } from "./messages"
-import { RoleGrantButtonError } from "./button"
-import { memberHasRole } from "../../../../../../utils/discord"
-import { parseHexColor } from "../../../../../../utils/color"
+import { RoleGrantModalError } from "../components/modal"
+import { ERR, MSG } from "../messages"
+import { RoleGrantButtonError } from "../components/button"
+import { memberHasRole } from "../../../../../../../utils/discord"
+import { parseHexColor } from "../../../../../../../utils/color"
 
 export const getModalCustomId = (interaction: Interaction, customId: string) => {
   const [prefix, guildId, channelId, roleId, buttonNameEnc, colorEnc] = customId.split(":")
